Autocreate on PROPFIND
If we don't have existing calendar folders then dav_parse_req_target
fails, because it looks for the base calendar via mupdate.
The new codepath simply attempts to lookup the inbox instead and then
proxies the request to that backend, triggering the autocreate.
This requires that the inbox has been created before though.